refactor: remove dead exports and stale re-exports
